% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMetadataToGRanges.R
\name{addMetadataToGRanges}
\alias{addMetadataToGRanges}
\title{Add metadata to a GenomicRanges object}
\usage{
addMetadataToGRanges(
  gr,
  df,
  gr_key = "gene_id",
  df_key = "gene_id",
  df_col_name = NA,
  rename_meta = NA
)
}
\value{
An annotated GenomicRanges object with a new metadata column added.
}
\description{
Function to add metadata columns to a GenomicRanges object. Potentially useful for \code{GenomicBreaks}.
}
