% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compDistance.R
\name{compDistance}
\alias{compDistance}
\title{Transform pair name into distance name.}
\usage{
compDistance(x, short = FALSE)
}
\arguments{
\item{x}{A vector of pair names such as \code{Oki_Kum}, \code{Osa_Nor}, etc}

\item{short}{Further group all Ciona and Drosophila together.}
}
\value{
A vector of "\verb{same_pop"}, \code{"Oki – North"} and \code{"North – North"}.
}
\description{
We have 3 main evolutionary distances: \emph{same population}, \emph{North Atlantic vs
North Pacific}, and \emph{Okinawa vs the other populations}.  To ease plotting
this function transforms the names of pairwise comparisons in one of these
three classes.
}
