% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compDistance.R
\name{compDistClass}
\alias{compDistClass}
\title{Transform pair name into distance class.}
\usage{
compDistClass(x)
}
\arguments{
\item{x}{A vector of pair names such as \code{Oki_Kum}, \code{Osa_Nor}, etc}
}
\value{
A vector of "\verb{same_pop"}, \code{"close"}, \code{"intermediate"}, and \code{distant}.
}
\description{
We have 4 main evolutionary distances: \emph{same population}, \emph{close},
\emph{intermediate}, \emph{distant}.  To ease plotting this function transforms the
names of pairwise comparisons in one of these three classes.
}
