% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compDistance.R
\name{compGenus}
\alias{compGenus}
\title{Transform pair name into genus name.}
\usage{
compGenus(x)
}
\arguments{
\item{x}{A vector of pair names such as \code{Oki_Kum}, \code{Osa_Nor}, etc}
}
\value{
A factor of \code{Oikopleura}, \code{Ciona}, and \code{Drosophila}, or \code{NA}.
}
\description{
We compare genomes within 3 genera: \emph{Oikopleura}, \emph{Ciona}, and \emph{Drosophila}.
To ease plotting this function transforms the names of pairwise comparisons
in one of these three names, or \code{NA} for comparisons between genera.
}
\details{
The levels are sorted so that \emph{Oikopleura} is at the top in default \code{ggplot}
figures, followed by \emph{Ciona} as both are tunicates.
}
