% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntenyTests.R
\name{isSynbrachial}
\alias{isSynbrachial}
\title{Flag pairs that are on homologous chromosome arms}
\usage{
isSynbrachial(gb)
}
\arguments{
\item{gb}{A \code{\link[GenomicBreaks:GBreaks-class]{GenomicBreaks::GBreaks}} object.}
}
\value{
A logical vector of same length as the \code{gb} object
}
\description{
This function assumes that homologous chromosomes have the same name, but
allows differences in character case.
}
