% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAllAnnotations.R
\name{loadAllAnnotations}
\alias{loadAllAnnotations}
\title{Load all annotations}
\usage{
loadAllAnnotations()
}
\value{
Returns a \code{\link{SimpleList}} of \code{TxDb} objects representing the
annotations of the genomes we study.
}
\description{
Accessory function to load all annotations provided by the \code{BreakpointsData}
package
}
