% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAlldNdS.R
\name{loadAlldNdS}
\alias{loadAlldNdS}
\title{Load dNdS data from OikScrambling data}
\usage{
loadAlldNdS(file, alignment_type = "GUIDANCE2")
}
\value{
Returns a \code{\link{SimpleList}} of data frames with dN/dS annotations.
}
\description{
Accessory function to load the dNdS table provided by the \code{BreakpointsData} package.
Requires \code{BreakpointsData} version >=3.6.1.
}
