% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAllRepeats.R
\name{loadAllRepeats}
\alias{loadAllRepeats}
\title{Load all repeat annotations}
\usage{
loadAllRepeats(compat = TRUE)
}
\arguments{
\item{compat}{Set to TRUE to keep the old names.  This option will be removed
when the migration will be complete.}
}
\value{
Returns a \code{SimpleList} of all loaded repeat annotations.
}
\description{
Load all the repeat annotations distributed by the \code{BreakpointsData} data
package, and sort them in broad classes by parsing their \code{Target} attribute.
}
