% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntenyTests.R
\name{nameSynbrachial}
\alias{nameSynbrachial}
\title{Reports syntenic chromosome arm name or NA}
\usage{
nameSynbrachial(gb)
}
\arguments{
\item{gb}{A \code{\link[GenomicBreaks:GBreaks-class]{GenomicBreaks::GBreaks}} object.}
}
\value{
A character vector of same length as the \code{gb} object
}
\description{
This function converts chromosome names to lower case.
}
