% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeUnplacedContigs.R
\name{removeUnplacedContigs}
\alias{removeUnplacedContigs}
\title{Remove unplaced contigs}
\usage{
removeUnplacedContigs(gb, target = TRUE, query = TRUE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{target}{Filter \emph{target} ranges.}

\item{query}{Filter \emph{query} ranges.}
}
\description{
Discard all rages whose sequence names do not match \emph{chr1}, \emph{Chr1}, \emph{chr2},
\emph{Chr2}, \emph{PAR}, \emph{XSR} or \emph{YSR}.
}
