% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scafs}
\alias{scafs}
\title{Scaffolding guides}
\format{
An object of class \code{SimpleList} of length 3.
}
\usage{
scafs
}
\description{
This object is generated by the vignettes \code{AomScaffolding.Rmd},
\code{KumeScaffolding.Rmd}, and \code{BergenScaffolding.Rmd}.  It contains \emph{guide} data
for the function \code{\link[GenomicBreaks:scaffoldByFlipAndMerge]{GenomicBreaks::scaffoldByFlipAndMerge}}.
}
\keyword{datasets}
