% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAllGenomes.R
\name{loadAllGenomes}
\alias{loadAllGenomes}
\title{Load all genomes}
\usage{
loadAllGenomes(compat = TRUE)
}
\arguments{
\item{compat}{Set to TRUE to keep the old names.  This option will be removed
when the migration will be complete.}
}
\value{
Returns a \code{\link{SimpleList}} of all loaded genomes.
}
\description{
Utility function to load all BSgenome packages in the active session and
reduce boilerplate in the vignettes.
}
