% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAllTranscriptGRs.R
\name{loadAllTranscriptsGR}
\alias{loadAllTranscriptsGR}
\title{Load all transcript information as GRanges}
\usage{
loadAllTranscriptsGR()
}
\value{
Returns a \code{\link{SimpleList}} of \code{GRanges} objects representing the
annotations of the genomes we study.
}
\description{
Accessory function to load all transcript annotations provided by the
\code{BreakpointsData} package
}
