% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OikScrambling.R
\docType{package}
\name{OikScrambling-package}
\alias{OikScrambling-package}
\title{The OikScrambling package}
\description{
The purpose of this package is to version-control and manage a bunch of
vignettes.  For this reason it does not provide functions.
}
\author{
The Luscombe laboratory
}
